Refactor code into multiple classes
