+++ src/main/java/seedu/lifeasier/commands/AddNotesCommand.java
@@ -1,6 +1,7 @@
 package seedu.lifeasier.commands;
 
 import seedu.lifeasier.storage.FileStorage;
+import seedu.lifeasier.storage.NoteStorage;
 import seedu.lifeasier.tasks.TaskList;
 import seedu.lifeasier.ui.Ui;
 import seedu.lifeasier.notes.Note;
@@ -25,6 +26,6 @@ public class AddNotesCommand extends Command {
         notes.add(new Note(noteTitle,noteDescription));
         System.out.println("Ok! I've taken note of this note!\n");
 
-        storage.writeToNoteSaveFile();
+        storage.saveNote();
     }
 }
